21 add service collection extension (#22)
* upgrade test packages

* add

* include di abstractions package

* add di tests

* add

* warn dont fail on TooManyRequests
