+++ .fossa.yml
@@ -1,0 +1,5 @@
+version: 3
+
+paths:
+  only:
+    - /CoinGeckoApi
