+++ Tests/CompaniesTests.cs
@@ -5,10 +5,17 @@
         [Test]
         public async Task GetCompaniesPublicTreasuryTest()
         {
-            var companiesResult = await Helpers.GetApiClient().Companies.GetCompaniesPublicTreasuryAsync();
+            try
+            {
+                var companiesResult = await Helpers.GetApiClient().Companies.GetCompaniesPublicTreasuryAsync();
 
-            Assert.That(companiesResult, Is.Not.Null);
-            Assert.That(companiesResult.Companies, Is.Not.Empty);
+                Assert.That(companiesResult, Is.Not.Null);
+                Assert.That(companiesResult.Companies, Is.Not.Empty);
+            }
+            catch (HttpRequestException ex) when (ex.StatusCode == System.Net.HttpStatusCode.TooManyRequests)
+            {
+                Assert.Warn(ex.Message);
+            }
         }
     }
 }
