+++ Tests/ExchangesTests.cs
@@ -5,62 +5,104 @@
         [Test]
         public async Task GetExchangesTest()
         {
-            var exchangesResult = await Helpers.GetApiClient().Exchanges.GetExchangesAsync();
+            try
+            {
+                var exchangesResult = await Helpers.GetApiClient().Exchanges.GetExchangesAsync();
 
-            Assert.That(exchangesResult, Is.Not.Null);
-            Assert.That(exchangesResult, Is.Not.Empty);
+                Assert.That(exchangesResult, Is.Not.Null);
+                Assert.That(exchangesResult, Is.Not.Empty);
+            }
+            catch (HttpRequestException ex) when (ex.StatusCode == System.Net.HttpStatusCode.TooManyRequests)
+            {
+                Assert.Warn(ex.Message);
+            }
         }
 
         [Test]
         public async Task GetExchangesListTest()
         {
-            var exchangesResult = await Helpers.GetApiClient().Exchanges.GetExchangesListAsync();
+            try
+            {
+                var exchangesResult = await Helpers.GetApiClient().Exchanges.GetExchangesListAsync();
 
-            Assert.That(exchangesResult, Is.Not.Null);
-            Assert.That(exchangesResult, Is.Not.Empty);
+                Assert.That(exchangesResult, Is.Not.Null);
+                Assert.That(exchangesResult, Is.Not.Empty);
 
-            var gdaxItem = exchangesResult.First(x => x.Id.Equals("bitstamp", StringComparison.InvariantCultureIgnoreCase));
+                var gdaxItem = exchangesResult.First(x => x.Id.Equals("bitstamp", StringComparison.InvariantCultureIgnoreCase));
 
-            Assert.That(gdaxItem, Is.Not.Null);
-            Assert.That(gdaxItem.Name, Is.EqualTo("Bitstamp"));
+                Assert.That(gdaxItem, Is.Not.Null);
+                Assert.That(gdaxItem.Name, Is.EqualTo("Bitstamp"));
+            }
+            catch (HttpRequestException ex) when (ex.StatusCode == System.Net.HttpStatusCode.TooManyRequests)
+            {
+                Assert.Warn(ex.Message);
+            }
         }
 
         [Test]
         public async Task GetExchangeTest()
         {
-            var exchangeResult = await Helpers.GetApiClient().Exchanges.GetExchangeAsync("gdax");
+            try
+            {
+                var exchangeResult = await Helpers.GetApiClient().Exchanges.GetExchangeAsync("gdax");
 
-            Assert.That(exchangeResult, Is.Not.Null);
-            Assert.That(exchangeResult.Tickers, Is.Not.Empty);
-            Assert.That(exchangeResult.Name, Is.EqualTo("Coinbase Exchange"));
+                Assert.That(exchangeResult, Is.Not.Null);
+                Assert.That(exchangeResult.Tickers, Is.Not.Empty);
+                Assert.That(exchangeResult.Name, Is.EqualTo("Coinbase Exchange"));
+            }
+            catch (HttpRequestException ex) when (ex.StatusCode == System.Net.HttpStatusCode.TooManyRequests)
+            {
+                Assert.Warn(ex.Message);
+            }
         }
 
         [Test]
         public async Task GetExchangeTickersTest()
         {
-            var tickersResult = await Helpers.GetApiClient().Exchanges.GetExchangeTickersAsync("gdax", new[] { "bitcoin", "ethereum", "cosmos" }, true);
+            try
+            {
+                var tickersResult = await Helpers.GetApiClient().Exchanges.GetExchangeTickersAsync("gdax", new[] { "bitcoin", "ethereum", "cosmos" }, true);
 
-            Assert.That(tickersResult, Is.Not.Null);
-            Assert.That(tickersResult.Tickers, Is.Not.Empty);
-            Assert.That(tickersResult.Name, Is.EqualTo("Coinbase Exchange"));
+                Assert.That(tickersResult, Is.Not.Null);
+                Assert.That(tickersResult.Tickers, Is.Not.Empty);
+                Assert.That(tickersResult.Name, Is.EqualTo("Coinbase Exchange"));
+            }
+            catch (HttpRequestException ex) when (ex.StatusCode == System.Net.HttpStatusCode.TooManyRequests)
+            {
+                Assert.Warn(ex.Message);
+            }
         }
 
         [Test]
         public async Task GetExchangeVolumeChartFriendlyTest()
         {
-            var chartResult = await Helpers.GetApiClient().Exchanges.GetExchangeVolumeChartFriendlyAsync("gdax", 2);
+            try
+            {
+                var chartResult = await Helpers.GetApiClient().Exchanges.GetExchangeVolumeChartFriendlyAsync("gdax", 2);
 
-            Assert.That(chartResult, Is.Not.Null);
-            Assert.That(chartResult, Is.Not.Empty);
+                Assert.That(chartResult, Is.Not.Null);
+                Assert.That(chartResult, Is.Not.Empty);
+            }
+            catch (HttpRequestException ex) when (ex.StatusCode == System.Net.HttpStatusCode.TooManyRequests)
+            {
+                Assert.Warn(ex.Message);
+            }
         }
 
         [Test]
         public async Task GetExchangeVolumeChartTest()
         {
-            var chartResult = await Helpers.GetApiClient().Exchanges.GetExchangeVolumeChartAsync("gdax", 2);
+            try
+            {
+                var chartResult = await Helpers.GetApiClient().Exchanges.GetExchangeVolumeChartAsync("gdax", 2);
 
-            Assert.That(chartResult, Is.Not.Null);
-            Assert.That(chartResult, Is.Not.Empty);
+                Assert.That(chartResult, Is.Not.Null);
+                Assert.That(chartResult, Is.Not.Empty);
+            }
+            catch (HttpRequestException ex) when (ex.StatusCode == System.Net.HttpStatusCode.TooManyRequests)
+            {
+                Assert.Warn(ex.Message);
+            }
         }
     }
 }
