+++ Tests/IndexesTests.cs
@@ -5,29 +5,50 @@
         [Test]
         public async Task GetIndexesTest()
         {
-            var indexesResult = await Helpers.GetApiClient().Indexes.GetIndexesAsync();
+            try
+            {
+                var indexesResult = await Helpers.GetApiClient().Indexes.GetIndexesAsync();
 
-            Assert.That(indexesResult, Is.Not.Null);
-            Assert.That(indexesResult, Is.Not.Empty);
+                Assert.That(indexesResult, Is.Not.Null);
+                Assert.That(indexesResult, Is.Not.Empty);
+            }
+            catch (HttpRequestException ex) when (ex.StatusCode == System.Net.HttpStatusCode.TooManyRequests)
+            {
+                Assert.Warn(ex.Message);
+            }
         }
 
         [Test]
         public async Task GetIndexTest()
         {
-            var indexResult = await Helpers.GetApiClient().Indexes.GetIndexAsync("cme_futures", "btc");
+            try
+            {
+                var indexResult = await Helpers.GetApiClient().Indexes.GetIndexAsync("cme_futures", "btc");
 
-            Assert.That(indexResult, Is.Not.Null);
-            Assert.That(indexResult.IsMultiAssetComposite, Is.False);
-            Assert.That(indexResult.Name, Is.EqualTo("CME Bitcoin Futures BTC"));
+                Assert.That(indexResult, Is.Not.Null);
+                Assert.That(indexResult.IsMultiAssetComposite, Is.False);
+                Assert.That(indexResult.Name, Is.EqualTo("CME Bitcoin Futures BTC"));
+            }
+            catch (HttpRequestException ex) when (ex.StatusCode == System.Net.HttpStatusCode.TooManyRequests)
+            {
+                Assert.Warn(ex.Message);
+            }
         }
 
         [Test]
         public async Task GetIndexesListTest()
         {
-            var indexesResult = await Helpers.GetApiClient().Indexes.GetIndexesListAsync();
+            try
+            {
+                var indexesResult = await Helpers.GetApiClient().Indexes.GetIndexesListAsync();
 
-            Assert.That(indexesResult, Is.Not.Null);
-            Assert.That(indexesResult, Is.Not.Empty);
+                Assert.That(indexesResult, Is.Not.Null);
+                Assert.That(indexesResult, Is.Not.Empty);
+            }
+            catch (HttpRequestException ex) when (ex.StatusCode == System.Net.HttpStatusCode.TooManyRequests)
+            {
+                Assert.Warn(ex.Message);
+            }
         }
     }
 }
