+++ Tests/Usings.cs
@@ -1,3 +1,4 @@
 global using CoinGeckoAPI;
 global using CoinGeckoAPI.Types;
+global using Microsoft.Extensions.DependencyInjection;
 global using NUnit.Framework;
